Add files via upload
This dataset is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset is to diagnostically predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset.
After conducting extensive Exploratory Data Analysis (EDA) and applying multiple machine learning models, we have gained significant insights into the dataset. Below are the key findings and conclusions: